refactor(statusline): move statusline defs to statusline_defs.h
